ticketstore/requirements.txt
Django==2.1.3
djangorestframework==3.9.0
pylint==2.1.1